requirements.txt
fastapi
uvicorn
tensorflow
tf_keras
pillow
numpy
python-multipart
